Move CvvInput classes to cvvinput package.
