+++ docs/com/checkout/android_sdk/Input/CvvInput.html
@@ -108,7 +108,7 @@
 <li><a href="../../../../com/checkout/android_sdk/Input/DefaultInput.html" title="class in com.checkout.sdk.input">com.checkout.sdk.input.DefaultInput</a></li>
 <li>
 <ul class="inheritance">
-<li>com.checkout.sdk.input.CvvInput</li>
+<li>com.checkout.sdk.cvvinput.CvvInput</li>
 </ul>
 </li>
 </ul>
