+++ sdk/src/main/res/layout/card_details.xml
@@ -159,7 +159,7 @@
                     app:errorEnabled="true"
                     app:errorTextAppearance="@style/Error">
 
-                    <com.checkout.sdk.input.CvvInput
+                    <com.checkout.sdk.cvvinput.CvvInput
                         android:id="@+id/cvv_input"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
